requirements.txt
streamlit
pygithub
pandas
plotly